backend/rec_quiz/views.py
from django.contrib.auth.models import User
from .models import Quiz
from .serializers import UserSerializer, QuizSerializer
from .permissions import IsOwnerOrReadOnly
from rest_framework import viewsets, permissions


class UserViewSet(viewsets.ReadOnlyModelViewSet):
    """
    This viewset automatically provides `list` and `detail` actions.
    """
    queryset = User.objects.all()
    serializer_class = UserSerializer


class QuizViewSet(viewsets.ModelViewSet):
    """
    This viewset automatically provides `list`, `create`, `retrieve`,
    `update` and `destroy` actions.
    """
    queryset = Quiz.objects.all()
    serializer_class = QuizSerializer
    permission_classes = (
        permissions.IsAuthenticatedOrReadOnly, IsOwnerOrReadOnly)

    def perform_create(self, serializer):
        serializer.save(owner=self.request.user)
