backend/requirements.txt
astroid==1.6.5
autopep8==1.4
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pycodestyle==2.4.0
pylint==1.9.3
pytz==2018.5
six==1.11.0
wrapt==1.10.11
